Minor updates to the text.
Specify 'internal sorting' at the top, and link to the wikipedia page.

Change 'like keys' to 'equal keys', which is probably easier for
non-native speakers to understand.

Attempt to hide away the big block on counting sort counter-size
relating to the range of keys. It's an aside, not super important,
and breaks the flow. Make it a footnote.

Link to @rygorous optimized histogramming code.